Bring back `--no-tests=warn` when needed `justfile`
This was temporarily removed to verify that we no longer need to
set `NEXTEST_NO_TESTS` to `fail` explicitly in `ci.yml` to get that
failing behavior, now that it is implied starting in
`cargo-nextest` version 0.9.85.

It is now brought back, so that we do not get that error in the one
place where we know there are no test cases yet do tell `nextest`
to run tests.