Have `check-mode.sh` handle errors better on macOS
In bash 3.2, which is used on macOS, `[[` commands don't invoke
`set -e` behavior. This adds an explicit check.

This also improves some code that sliced an array instead of using
the variables its elements were already assigned to. One of them is
used twice, and indexing in the existing usage would make the
meaning less clear. So the variables would not be reasonable to
eliminate by slicing the array, without other refactoring.