Use consistent `git` command name in gix-testtools
In the test tools, this always runs the `git` program as `git.exe`
on Windows, while continuing always to run it as `git` on other
systems.

Prior to this change, on Windows `gix-testtools` used `git` in some
operations and `git.exe` in others:

- `parse_git_version` used `git.exe`.
- Other functions used `git`.
- For the git daemon, `git-daemon.exe` was used.

For the way `gix-testtools` uses the `git` program, it would be
fine to call it `git` on all platforms. For example, it does not
form full paths to the executable that have to be found to exist
in operations other than running it. (For running it, the `.exe`
suffix is allowed to be omitted.) So it would probably be fine to
use the even simpler logic of having it be `git` everywhere. But
since `git.exe` was sometimes used, `git-daemon.exe` was used, and
using `git.exe` is more similar to the behavior in `git-path`, it
is changed to use `git.exe` when the platform is Windows.

Because `gix-testtools` does not depend on `gix-path`, it doesn't
use `gix_path::env::exe_invocation` to decide how to call `git`.
That keeps it from finding `git` in some Windows environments
(those where `git` is in a standard/predictable location but not in
`PATH`). This change has no effect on that limitation.