Impl topo::Builder::from_iters with new, with_tips and with_ends
Previously, the only way to create a `Builder` would be via
`Builder::from_iters`. That fn takes as arguments both the tips and
ends, and used to be the only possibility for specifying either of them
during the building process. However, in order to enhance the builder-
likeness of `Builder`, we recently introduced various fns allowing for
alternative build flows.

This fn reduces code duplication by using those new fns for implementing
`from_iters`.