trim `gix log` to be mergeable.
For now, let's ignore file-based logs as it would take a moment to make it fast enough.
Git can diff with pathspecs, which is probably something that would be needed to be fast here.
In any case, more research would be needed to be competitive in performance.

Maybe one day this will be re-added even in its current form to help with `gix blame` debugging,
but I'd hope that there will be better ways to validate it.