Split out ARM64 test-fast from matrix, make nonrequired
As suggested in:
https://github.com/GitoxideLabs/gitoxide/issues/1790#issuecomment-2606481883

It likely won't have to be kept this way. But making it nonrequired
for now makes it so that investigating what triggers the SIGSEGV
(and SIGBUS) errors -- as well as other errors that were found
while investigating that (d9e7fdb, e71b0cf, 5a71963) -- doesn't
have to be rushed.