Redo experiment 1 with current beta and nightly
When using `dtolnay/rust-toolchain` with the `toolchain` key to
specify a channel, the action version should be given as `@master`.
But I accidentally kept it at `@stable`! This caused `beta` and
`nightly` to refer to the most recent beta and nightly builds
*prior* to the current stable version. That made the conclucions
about beta and nightly builds inaccurate. This rectifies that
error and repeats the experiment.

See e71b0cf (1f3f6b5), #1790, and rust-lang/rust#135867 for context.

(I made this mistake in both experiment 1 and experiment 2, having
wrongly thought I'd changed `@stable` to `@master` for experiment
1. This commit just repeats experiment 1, but experiment 2 should
also be repeated for the same reason.)