doc: Clarify description of `single` escaping
This improves how `gix_quote::single` documents how it handles the
cases it treats specially.

The fix in #717 (discussed in #713) wasn't accompanied by a change
to the documentation, which continued either to describe the
preceding behavior or to be ambiguous. This changes it to describe
the current behavior unambiguously. Only documentation is changed.