Drop unused "Unicode-DFS-2016" license from allowed list
`cargo deny` warned that "Unicode-DFS-2016" appeared in the list of
allowed licenses but was not used by any dependency. This list had
both "Unicode-3.0" and "Unicode-DFS-2016".

Possibly relevant is that `unicode-ident` updated its license
in 1.0.14 (https://github.com/dtolnay/unicode-ident/pull/28),
changing from "Unicode-DFS-2016" to "Unicode-3.0".

The "Unicode-3.0" license was added to `deny.toml` in 5173e9a
(#1668), but "Unicode-DFS-2016" was not removed at that time. It
looks like some Unicode-related packages still used that license at
that time. But it does not seem to be needed here anymore.