Put back ubuntu-24.04-arm and test heavily again
Since the known issue looks like it has been fixed by switching the
runners to different virtual machines.

This repeats experiment 3 as well as doing a comparable experiment
for the 32-bit containerized test, which had also failed to run
some commands in Docker, where it was unclear if it was somehow
related or a separate problem.