Move `shell()` helpers to a helper module
The new `auxiliary` module within `gix_path::env` is a sibling of
the `git` module and is similarly an implementation detail only.

So far the only "auxiliary" program this module finds is `sh`.

The module is named `auxiliary` rather than `aux` because Windows
has problems with files named like `aux` or `aux.rs`, due to `AUX`
being a reserved device name.