doc: Simplify commit-graph doc link
The `#:~:text` in the anchor was no longer working, as tested in
Chromium-based browsers that ordinarily support such links. Like
any broken anchor, it didn't keep the page from resolving. But it
didn't jump to (nor participate in any highlighting of) specific
text.

Fortunately, the relevant text is near the top of the page. So the
original intent of the longer and more complicated URL is satisfied
in the simpler URL.