Improve filesystem timestamp granularity check
This avoids false positives for the check of whether we can proceed
to run the filesystem snapshot `journey` test, by making a hundred
files about as fast as we can using ordinary non-parallelized
techniques, then checking if all their timestamps are distinct.

This carries a small risk of false negatives, but experiments on
macOS, where the `journey` test is typically able to run, suggest
that the rate of false negatives will be low. (False positive are
also, in theory, possible, but should be extremely rare or never
happen.)