Refactor `summarize` recipe to be easier to read
Most recipes in `justfile` that use `cargo nextest run` write it
out in a command, even though a `nextest` recipe exists. This is
often done because the command appears in a non-leading line of the
recipe, so having the recipe depend on the `nextest` recipe would
not be feasible. However, it also has the benefit of being more
clear about exactly what command is being run, especially when the
arguments are long and complicated, as in `summarize`.

This changes `summarize` so that it invokes nextest explicitly as a
step of the recipe, rather than depending on `nextest` and passing
its arguments through.