Fix clippy
This removes some extra unnecessary semicolons that `clippy` has
recently begun to catch, causing CI to fail with errors such as:

    error: unnecessary semicolon
      --> gix-transport/src/client/blocking_io/http/traits.rs:30:18
       |
    30 |                 };
       |                  ^ help: remove
       |
       = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_semicolon
       = note: `-D clippy::unnecessary-semicolon` implied by `-D warnings`
       = help: to override `-D warnings` add `#[allow(clippy::unnecessary_semicolon)]`

While it looks like this might first have been observed in #1917,
it is unrelated to any change there. It happens when the current
tip of main (4660f7a) is rerun, as observed in:
https://github.com/EliahKagan/gitoxide/actions/runs/14254079128/job/39958292846

This runs `just clippy-fix` and `etc/copy-packetline.sh` to fix it.